Refactored UnZip to ZipUtil and moved to core to share it with NodejsInstall
