+++ core/tern.core.tests/src/tern/server/nodejs/process/PathHelper.java
@@ -13,6 +13,8 @@ package tern.server.nodejs.process;
 import java.io.File;
 import java.io.IOException;
 
+import tern.utils.UnZip;
+
 public class PathHelper {
 
 	/**
