+++ eclipse/tern.eclipse.ide.server.nodejs.core/src/tern/eclipse/ide/server/nodejs/internal/core/NodejsInstall.java
@@ -13,15 +13,12 @@ package tern.eclipse.ide.server.nodejs.internal.core;
 import java.io.File;
 import java.io.IOException;
 
-import org.apache.tools.ant.taskdefs.Expand;
-import org.apache.tools.ant.taskdefs.GUnzip;
-import org.apache.tools.ant.taskdefs.Unpack;
-import org.apache.tools.ant.taskdefs.Untar;
 import org.eclipse.core.runtime.FileLocator;
 import org.eclipse.core.runtime.IConfigurationElement;
 import org.eclipse.core.runtime.Platform;
 
 import tern.eclipse.ide.server.nodejs.core.INodejsInstall;
+import tern.utils.UnZip;
 
 public class NodejsInstall implements INodejsInstall {
 
@@ -33,7 +30,7 @@ public class NodejsInstall implements INodejsInstall {
 
 	/**
 	 * GeneratorType constructor comment.
-	 * 
+	 *
 	 * @param element
 	 *            a configuration element
 	 * @throws IOException
@@ -47,21 +44,18 @@ public class NodejsInstall implements INodejsInstall {
 			File baseDir = FileLocator.getBundleFile(Platform
 					.getBundle(pluginId));
 			this.path = new File(baseDir, path);
-			
+
 			// check if path exists, if it doesn't look for zip
 			if (!this.path.exists()) {
 				String zip = element.getAttribute("zip");
-				
+
 				File zipFile = new File(baseDir, zip);
-				
+
 				if (zipFile.exists()) {
 					if (zipFile.getName().toLowerCase().endsWith(".zip")) {
-						Expand unzip = new Expand();
-						unzip.setDest(baseDir);
-						unzip.setSrc(zipFile);
-						unzip.execute();
+						UnZip.extract(zipFile, baseDir);
 					}
-					
+
 					if(this.path.exists()) {
 						this.path.setExecutable(true);
 					}
